+++ tests/app/server.js
@@ -1913,7 +1913,15 @@ app.get('/svelte/frame-remember', (req, res) =>
   inertia.render(req, res, { component: 'Svelte/FrameRemember', props: {} }),
 )
 
+app.get('/svelte/frame-remember-skip', (req, res) =>
+  inertia.render(req, res, { component: 'Svelte/FrameRememberSkip', props: {} }),
+)
+
 app.get('/svelte/frame-remember/pane', (req, res) =>
+  inertia.render(req, res, { component: 'Svelte/FrameRememberPane', props: {} }),
+)
+
+app.get('/svelte/frame-remember/pane-skip', (req, res) =>
   inertia.render(req, res, { component: 'Svelte/FrameRememberPane', props: {} }),
 )